Document and test object cloning
