+++ tests/Unit/CrudUnitTest.php
@@ -8,34 +8,12 @@
 namespace Alley\WP\Big_Pit\Tests\Unit;
 
 use Alley\WP\Big_Pit;
-use PHPUnit\Framework\TestCase;
+use Alley\WP\Big_Pit\Tests\TestCase;
 
 /**
  * CRUD tests.
  */
 class CrudUnitTest extends TestCase {
-	/**
-	 * Create the database table.
-	 */
-	public static function setUpBeforeClass(): void {
-		parent::setUpBeforeClass();
-
-		Big_Pit::instance()->boot();
-	}
-
-	/**
-	 * Drop the database table.
-	 */
-	public static function tearDownAfterClass(): void {
-		global $wpdb;
-
-		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery, WordPress.DB.DirectDatabaseQuery.NoCaching, WordPress.DB.DirectDatabaseQuery.SchemaChange
-		$wpdb->query( 'DROP TABLE ' . $wpdb->big_pit );
-		delete_option( 'wp_big_pit_database_version' );
-
-		parent::tearDownAfterClass();
-	}
-
 	/**
 	 * Test CRUD operations.
 	 */
@@ -67,50 +45,5 @@ class CrudUnitTest extends TestCase {
 
 		// key3 should still be there.
 		$this->assertSame( $val3, $big_pit->get( $key3, $grp2 ) );
-	}
-
-	/**
-	 * Test for expected number of queries with the in-memory cache.
-	 */
-	public function test_in_memory_cache() {
-		global $wpdb;
-
-		$big_pit = Big_Pit::instance();
-
-		// Two fetches of the same key should only result in one query.
-		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
-		$big_pit->get( 'key1', 'group1' );
-		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
-
-		$big_pit->set( 'key1', 'value1', 'group1' );
-
-		// Value has changed, so there should be another query.
-		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
-		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
-
-		// Fetching it again should not result in another query.
-		$big_pit->get( 'key1', 'group1' );
-		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
-
-		$big_pit->delete( 'key1', 'group1' );
-
-		// Value has changed, so there should be another query.
-		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
-		$this->assertSame( 1, $wpdb->num_queries - $num_queries_before );
-
-		$big_pit->set( 'key1', 'value1', 'group1' );
-		$big_pit->set( 'key2', 'value2', 'group1' );
-		$big_pit->get( 'key1', 'group1' );
-		$big_pit->get( 'key2', 'group1' );
-		$big_pit->flush_group( 'group1' );
-
-		// All values have changed, so there should be queries for each value in the group that was set.
-		$num_queries_before = $wpdb->num_queries;
-		$big_pit->get( 'key1', 'group1' );
-		$big_pit->get( 'key2', 'group1' );
-		$this->assertSame( 2, $wpdb->num_queries - $num_queries_before );
 	}
 }
